[3] Als Agent möchte ich, dass die unbeantwortete Kundenliste im Livechat von Alt nach Neu sortiert ist
Introduced new persistent property to capture the timestamp which is the oldest one issued by a customer with no response by an agent afterwards.
If an agent answered lateron, the timestamp of the last message by the customer prior to the agent's answer is stored in oder to always have a relevant timestamp recorded.